backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
itsdangerous
sqlalchemy
psycopg2-binary
pyyaml
